% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_completion.R
\name{chatgpt_chat_completion}
\alias{chatgpt_chat_completion}
\title{Generate chat completion using OpenAI GPT large language models}
\usage{
chatgpt_chat_completion(messages, model = "gpt-3.5-turbo", token = NULL)
}
\arguments{
\item{messages}{A character vector or list of messages exchanged in the chat. Each message should be a list containing 'role' and 'content' keys. The 'role' can be 'system', 'user', or 'assistant'. The 'content' contains the text of the message.}

\item{model}{The model to be used for chat generation. Defaults to "gpt-3.5-turbo".}

\item{token}{Your OpenAI API key. If not provided, the function will attempt to retrieve it from the environment variable OPENAI_API_KEY.}
}
\value{
The generated completion text.
}
\description{
This function generates a chat completion using OpenAI large language models.
}
\examples{
chatgpt_chat_completion("What is the meaning of life?")
}
