% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_completion.R
\name{chatgpt_completion}
\alias{chatgpt_completion}
\title{Generate completion with OpenAI}
\usage{
chatgpt_completion(prompt, model = "gpt-3.5-turbo", token = NULL)
}
\arguments{
\item{prompt}{The text prompt to generate completion for.}

\item{model}{The model to use for completion. Defaults to "gpt-3.5-turbo".}

\item{token}{An optional API key. If not provided, the function will attempt to retrieve the key from the environment variable "OPENAI_API_KEY".}
}
\value{
The generated completion text.
}
\description{
This function sends a request to the OpenAI API to generate a completion for a given prompt.
}
\examples{
# note; currently returns 404?
# chatgpt_completion("Once upon a time")
}
