% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_completion.R
\name{explain_data}
\alias{explain_data}
\title{explain_data}
\usage{
explain_data(
  data,
  subject,
  audience = "a faculty board",
  user_message_template = c("Describe {subject} for {audience}.\\n\\n{table}") \%>\%
    stringr::str_flatten(" "),
  system_message_template = c("You are a data scientist providing context and",
    "insight on various data visualizations.",
    "Your responses should be concise and to the point,",
    "and targeted at the requested audience.", "You should highlight common patterns,",
    "interactions between variables,", "and point out discrepancies in the data.",
    "You are allowed to use basic markdown markup,",
    "but should not add headers, lists, etc.") \%>\% stringr::str_flatten(" "),
  token = NULL,
  model = "gpt-4"
)
}
\arguments{
\item{data}{A data frame containing the data set to be explained.}

\item{subject}{A character string specifying the subject of the explanation.}

\item{audience}{A character string specifying the intended audience for the explanation (default is "a faculty board").}

\item{user_message_template}{A character string template for the user's message to the chatbot. The template can contain placeholders {subject} and {audience} to be filled with the actual values. The default template includes a placeholder for {table}, which will be replaced with a formatted table of the data set.}

\item{system_message_template}{A character string template for the system's message to the chatbot. This message provides guidance to the model in generating the explanation. The default message is probably fine for most cases.}

\item{token}{An optional authentication token for the chatgpt_chat_completion function (default is NULL).}

\item{model}{A character string specifying the GPT model to be used for chatbot completion (default is "gpt-4").}
}
\value{
This function returns the response generated by the chatgpt_chat_completion function.
}
\description{
This function generates an explanation of a specific subject for a specific audience, given the underlying dataset.
}
\examples{
# Generate an explanation for a faculty board
explain_data(
  data = mtcars,
  subject = "Figure 1: engine performance and properties of various cars.",
  audience = "a faculty board")

# Generate an explanation for a different audience
explain_data(
  data = mtcars,
  subject = "Figure 1: engine performance and properties of various cars.",
  audience = "a group of toddlers")
}
